Add LinkGenerator class
The LinkGenerator class is the base for issue #10. Models that inherit
from it will need to define methods that return the url for the view,
edit and delete endpoints. This was not done as abstract methods because
there were issues with the multi inheritance of the db model classes.

the next step is letting classes derive from it and to replace all the
manual linking with the functions. this will hopefully reduce code
clutter and add consistency to linkage. it will also aid with the
impending removal of glyphicons, as there is now a central point where
button icons are defined (with the exception of custom icon parameters).

Signed-off-by: Thorben Römer <thorbenroemer@t-online.de>