Media: add asynchronous upload form to editor
The editor now has a new button, which opens a modal dialog to upload a
new file asynchronously. After a successful upload, the file can be
directly integrated into the editor window.

The new javascript class MediaUploader does all the work and calls a
callback function after a successful upload, containing information
about the new media file. This approach allows for easy re-use of this
class for things like profile-pictures. It could even be extended to be
a generic async-form handler with little effort.

Signed-off-by: Thorben Römer <thorbenroemer@t-online.de>