Implement new Media Sidebar (#49)
The Markdown Editor now has a new "Sidebar" which is actually a Bootstrap Modal. It provides a thumbnail preview for images, has multi select and its own upload modal.

It uses an extendable javascript base class, that will be used in the future to update the other sidebars, after some feedback has been gathered.