Session: fix date for session.edit
convert utc date to local date for datepicker. note that we use the
server-side value instead of the client value (of the form field). this
way, refreshing the form does not apply the correction again, which
would result in a different value on every refresh. as a downside, the
value is obviously not cachable by the browser.

Signed-off-by: Thorben Römer <thorbenroemer@t-online.de>