Multi: associate Parties with Campaigns
link up campaigns and parties via the Campaign form. associated parties
are displayed on the campaign page and associated campaigns are
displayed on the party page. this now allows to check if a player is a
DM of a 'party' by association with the Campaign, thus allowing DMs to
edit the party page.

further improvements:
- let the DM also change party members
- hook up some js so that default_participants will be automatically
selected/deselected if you associate a party (probably with a popup if
this action is wanted)

Signed-off-by: Thorben Römer <thorbenroemer@t-online.de>